ADD : pathService to add Path DB and search
Former-commit-id: 3ec4112734cbc0cdf992c9829e1b61671a705589

+++ BE/pijja/src/main/java/com/A605/pijja/domain/plan/entity/Path.java
@@ -23,6 +23,6 @@ public class Path {
 
     private float distance;
 
-    private Long time;
+    private float time;
 
 }
